EmptyTocd moved from kapua-certificate-internal to kapua-commons
